+++ lib/frp/src/debug.rs
@@ -1,0 +1,176 @@
+
+
+use crate::prelude::*;
+use crate::MessageType;
+
+// ================
+// === Graphviz ===
+// ================
+
+/// Visualization data for a nodes.
+#[derive(Debug,Clone)]
+pub struct VizNode {
+    display_id : usize,
+    variant    : String,
+    label      : String,
+}
+
+impl VizNode {
+    /// Constructor
+    pub fn new(display_id:usize, variant:String, label:String) -> Self {
+        VizNode {display_id,variant,label}
+    }
+}
+
+
+/// Visualization data for a link between nodes.
+#[derive(Debug,Clone)]
+pub struct VizLink {
+    source_display_id : usize,
+    target_display_id : usize,
+    message_type      : MessageType,
+    data_type         : String,
+}
+
+impl VizLink {
+    /// Constructor.
+    pub fn new
+    (source_display_id:usize, target_display_id:usize, message_type:MessageType, data_type:String)
+     -> Self {
+        Self {source_display_id,target_display_id,message_type,data_type}
+    }
+}
+
+
+/// Graphviz FRP system visualizer.
+#[derive(Debug,Default)]
+pub struct Graphviz {
+    nodes  : HashMap<usize,VizNode>,
+    labels : HashMap<usize,String>,
+    links  : Vec<VizLink>,
+}
+
+impl Graphviz {
+    /// Defines a new node.
+    pub fn add_node<Tp:Str,Label:Str>
+    (&mut self, id:usize, display_id:usize, tp:Tp, label:Label) {
+        let tp    = tp.into();
+        let label = label.into();
+        self.nodes.insert(id,VizNode::new(display_id,tp,label.clone()));
+        self.labels.insert(id,label);
+    }
+
+    /// Defines a new link between nodes.
+    pub fn add_link<DataType:Str>
+    (&mut self, source:usize, target:usize, message_type:MessageType, data_type:DataType) {
+        let link = VizLink::new(source,target,message_type,data_type.into());
+        self.links.push(link);
+    }
+
+    /// Checks if a node with the given id is already registered in the node map.
+    pub fn contains(&mut self, id:usize) -> bool {
+        self.nodes.contains_key(&id)
+    }
+
+    /// Takes a set of nodes and outputs a map from `display_id` to a particular node. In case the
+    /// `display_id` points to several nodes, the node types `Hold` and `Recursive` has weaker
+    /// preference.
+    fn create_node_map(&self) -> HashMap<usize,VizNode> {
+        let mut node_map : HashMap<usize,VizNode> = default();
+        for (_,node) in &self.nodes {
+            let entry        = node_map.entry(node.display_id);
+            let merged_entry = entry.and_modify(|node2|{
+                let variant = &node2.variant;
+                if variant == "Hold" || variant == "Recursive" {
+                    *node2 = node.clone();
+                }
+            });
+            merged_entry.or_insert(node.clone());
+        }
+        node_map
+    }
+
+    /// Outputs a Graphviz Dot code.
+    pub fn to_code(&self) -> String {
+        let mut code = String::default();
+        let node_map = self.create_node_map();
+
+        for (_,node) in &node_map {
+            let color = match &node.variant as &str {
+                "Toggle"  => "534666",
+                "Gate"    => "e69d45",
+                "Hold"    => "308695",
+                "Lambda"  => "d45769",
+                "Lambda2" => "d45769",
+                _         => "455054",
+            };
+            let fill    = iformat!("[fillcolor=\"#{color}\"]");
+            let spacing = "<br/><FONT POINT-SIZE=\"5\"> </FONT><br/>";
+            let variant = iformat!("<FONT POINT-SIZE=\"9\">{node.variant}</FONT>");
+            let label   = iformat!("[label=< {node.label} {spacing} {variant} >]");
+            let line    = iformat!("\n{node.display_id} {fill} {label}");
+            code.push_str(&line);
+        }
+
+        for link in &self.links {
+            let source    = &link.source_display_id;
+            let target    = &link.target_display_id;
+            let data_type = &link.data_type;
+            let not_loop  = source != target;
+            if not_loop {
+                let style = match link.message_type {
+                    MessageType::Behavior => "[style=\"dashed\"]",
+                    _ => ""
+                };
+                let label = if data_type == "()" { "" } else { &data_type };
+                let label = iformat!("[label=\"  {label}\"]");
+                let line  = iformat!("\n{source} -> {target} {style} {label}");
+                code.push_str(&line);
+            }
+        }
+
+        let fonts      = "[fontname=\"Helvetica Neue\" fontsize=11]";
+        let node_shape = "[shape=box penwidth=0 margin=0.12 style=\"rounded,filled\"]";
+        let node_style = "[fontcolor=white fillcolor=\"#5397dc\"]";
+        let edge_style = "[arrowsize=.7 fontcolor=\"#555555\"]";
+        let graph_cfg  = iformat!("rankdir=TD; graph {fonts};");
+        let nodes_cfg  = iformat!("node {fonts} {node_shape} {node_style};");
+        let edges_cfg  = iformat!("edge {fonts} {edge_style};");
+        iformat!("digraph G {{ \n{graph_cfg} \n{nodes_cfg} \n{edges_cfg} \n{code} \n}}")
+    }
+}
+
+impl From<Graphviz> for String {
+    fn from(cfg:Graphviz) -> String {
+        cfg.to_code()
+    }
+}
+
+
+
+// =======================
+// === GraphvizBuilder ===
+// =======================
+
+pub trait GraphvizBuilder {
+    fn graphviz_build(&self, builder:&mut Graphviz);
+
+    fn to_graphviz(&self) -> String {
+        let mut builder = Graphviz::default();
+        self.graphviz_build(&mut builder);
+        builder.into()
+    }
+
+    fn display_graphviz(&self) {
+        let code = self.to_graphviz();
+        let url  = percent_encoding::utf8_percent_encode(&code,percent_encoding::NON_ALPHANUMERIC);
+        let url  = format!("https://dreampuf.github.io/GraphvizOnline/#{}",url);
+        crate::web::window().open_with_url_and_target(&url,"_blank").unwrap();
+    }
+}
+
+impl<T:?Sized+GraphvizBuilder> GraphvizBuilder for Rc<T> {
+    fn graphviz_build(&self, builder:&mut Graphviz) {
+        self.deref().graphviz_build(builder)
+    }
+}
